refactor: rewrite and rename cjsToEsm.cjs to ECMAScript notation
